refactor(extensions): move OpenCode config projection into adapter
