+++ src/crates/adapters/opencode-adapter/AGENTS.md
@@ -6,7 +6,8 @@ The current crate owns OpenCode user Instruction path/config precedence, the sta
 managed-package path, the OpenCode-specific implementations of command,
 standalone-tool, subagent, and MCP provider contracts, the bounded projection of
 configured local Skill roots, the private merged configuration snapshot passed
-to the managed Plugin Host's Config Hook, and runtime-free mapping of
+to the managed Plugin Host's Config Hook, the typed Agent/permission/Tool/Skill
+projection of that Hook's result, and runtime-free mapping of
 caller-normalized static-preview Hook descriptors. It preserves OpenCode source
 discovery, precedence, formats, argument expansion, and versioned compatibility semantics.
 Shared source catalog, lifecycle coordination, file-watch implementation,
@@ -41,8 +42,9 @@ Product-source boundary:
   replacement. This adapter supplies OpenCode-qualified source identity/order and
   watch roots through narrow provider contracts; the reusable file-watch service
   supplies change facts. For the current package-plugin path, Core owns logical
-  workspace generations and capability publication while this adapter owns the
-  typed Host wire/connection and uses services process-tree primitives. The legacy
+  workspace generations and capability publication, this crate owns the typed
+  Config projection, and the sibling `opencode-plugin-host` crate owns the typed
+  Host wire/connection while using services process-tree primitives. The legacy
   managed-package path keeps request reliability in `PluginRuntimeClient`, and
   standalone workers remain under `ScriptToolRuntime`; do not merge these lifecycle facts.
 - Effective policy and safe-start mode must be recomputed before third-party
@@ -80,6 +82,10 @@ Product-source boundary:
   Config Hook, including unknown OpenCode fields. Inline content is bounded,
   uses a redacted virtual source identity, has no watch root, and may resolve
   relative paths only from an explicit workspace context.
+- Managed Plugin Host Config projection validates contributor attribution and
+  maps only the currently consumed Agent, permission, Plugin Tool, and workspace
+  Skill fields. It must not create Agent runtime keys, select native Tool
+  baselines, mutate product registries, or grow into a generic Config platform.
 - Current source inspection recognizes only the tested declarative subset. The
   adapter may reuse the workspace-pinned parse-only OXC profile for syntax-safe static
   projection, but it is not a general JavaScript/TypeScript semantic analyzer or
@@ -106,8 +112,8 @@ Product-source boundary:
   import the adapter directly. The composition layer does not
   discover dynamic sources, prepare dependencies, or import plugin modules.
 - Product Assembly may consume this crate only from reviewed composition modules
-  such as `bitfun-core/plugin_runtime`, `bitfun-core/external_sources`, or
-  `bitfun-core/instruction_sources`; boundary
+  such as `bitfun-core/plugin_runtime`, `bitfun-core/external_sources`,
+  `bitfun-core/instruction_sources`, or `bitfun-core/plugin_config_publication`; boundary
   guards and focused assembly-path tests must change with any additional consumer.
 - This crate must not depend on Codex, Claude Code, or another ecosystem adapter.
   New ecosystems are sibling adapters registered by Product Assembly, not modes of
@@ -124,5 +130,6 @@ Product-source boundary:
 - `cargo test -p bitfun-opencode-adapter --test opencode_static_source_contracts opencode_subagent_adapter::`
 - `cargo test -p bitfun-opencode-adapter --test tool_source_contracts`
 - `cargo test -p bitfun-opencode-adapter --lib p0_c2_fixture`
+- `cargo test -p bitfun-opencode-adapter --lib plugin_config_projection`
 - `cargo test -p bitfun-opencode-adapter --lib client_path_projects_trusted_custom_tool_candidate_with_permission_prompt`
 - `node scripts/check-core-boundaries.mjs`
