+++ src/crates/assembly/core/src/lib.rs
@@ -37,7 +37,7 @@ pub mod native_hooks;
 #[cfg(all(test, feature = "agent-runtime"))]
 mod native_hooks_tests;
 #[cfg(feature = "opencode-plugin-host")]
-mod plugin_config_projection;
+mod plugin_config_publication;
 #[cfg(feature = "opencode-plugin-host")]
 mod plugin_hook_bridge;
 #[cfg(feature = "opencode-plugin-host")]
